requirements.txt
aiofiles
asgiref
click
fastapi==0.95.2
uvicorn==0.22.0
h11
Jinja2
MarkupSafe
pydantic>=1.8,<2.0
starlette==0.27.0
typing-extensions
kafka-python-ng
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt==3.2.2

sqlalchemy
python-multipart
minio==7.1.14
